requirements.txt

hadoop=3.4.2
hive=3.1.3
spark=3.5.8
pandas
wsl
postgresql=18.0
mongodb
git